requirements.txt
coverage==6.2
Django==3.2.21
djangorestframework==3.14.0
itsdangerous==2.0.1
Pillow==8.4.0
whitenoise==5.3.0
